refactor: 각 경로에 대한 정보는 Station이 갖고 있도록 수정
