+++ README.md
@@ -12,6 +12,7 @@
 [경로 관련 기능]
 - [x] 각 경로는 거리, 시간 정보를 담고 있어야 한다
 - [x] 노선에 경로를 추가할 수 있어야 한다
+- [ ] 각 경로(역)가 추가될 때 이전 역에서 추가될 역 정보를 업데이트 해야함 
 
 [경로 조회 기능]
 - [ ] 출발역과 도착열을 입력받아 경로를 조회 할 수 있어야 한다
