+++ src/main/java/subway/domain/line/Line.java
@@ -1,20 +1,16 @@
 package subway.domain.line;
 
-import subway.domain.path.Path;
+import subway.domain.station.Station;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Line {
-    private List<Path> paths = new ArrayList<>();
+    private List<Station> stations = new ArrayList<>();
     private String name;
 
     public Line(String name) {
         this.name = name;
-    }
-
-    public void addPath(Path path) {
-        paths.add(path);
     }
 
     public String getName() {
