+++ src/main/java/subway/domain/station/Station.java
@@ -1,7 +1,11 @@
 package subway.domain.station;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class Station {
     private String name;
+    private List<Path> paths = new ArrayList<>();
 
     public Station(String name) {
         this.name = name;
@@ -11,5 +15,7 @@ public class Station {
         return name;
     }
 
-    // 추가 기능 구현
+    public void addPath(Path path) {
+        paths.add(path);
+    }
 }
